server/requirements.txt
fastmcp>=2.0.0
starlette>=0.27.0
uvicorn>=0.23.0
docker>=7.0.0
pyyaml>=6.0
watchdog>=3.0.0
